# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing TV series catalog",
        "title": "Serial Catalog API",
        "contact": {
            "email": "evgenij.bx@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/otus.serial.add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new serial to the catalog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serials"
                ],
                "summary": "Create new TV serial",
                "parameters": [
                    {
                        "description": "Serial data",
                        "name": "serial",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/catalog.Serial"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SerialItemResponse"
                        }
                    }
                }
            }
        },
        "/otus.serial.delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a TV serial from catalog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serials"
                ],
                "summary": "Delete serial",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Serial ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/otus.serial.get/{id}": {
            "get": {
                "description": "Get detailed information about a TV serial",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serials"
                ],
                "summary": "Get serial by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Serial ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved serial",
                        "schema": {
                            "$ref": "#/definitions/handler.SerialItemResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/otus.serial.update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update existing TV serial",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "serials"
                ],
                "summary": "Update serial",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Serial ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "serial",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/catalog.Serial"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SerialItemResponse"
                        }
                    },
                    "400": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "catalog.Serial": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "TV series"
                },
                "duration": {
                    "type": "number",
                    "example": 40
                },
                "file_id": {
                    "type": "integer",
                    "example": 0
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "production_period": {
                    "type": "string",
                    "example": "2008-2013"
                },
                "quality": {
                    "type": "string",
                    "example": "High"
                },
                "rating": {
                    "type": "number",
                    "example": 9.5
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Breaking Bad"
                }
            }
        },
...
        "handler.DeleteResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "deleted": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "entity not found"
                }
            }
        },
...
        "handler.SerialItemResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "item": {
                            "$ref": "#/definitions/catalog.Serial"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}